% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr_1d.R
\name{get_hdr_1d}
\alias{get_hdr_1d}
\title{Computing the highest density regions of a 2D density}
\usage{
get_hdr_1d(
  method = "kde",
  x,
  probs = c(0.99, 0.95, 0.8, 0.5),
  n = 512,
  range,
  parameters = list(),
  HDR_membership = TRUE,
  fun,
  args = list()
)
}
\arguments{
\item{method}{temp}

\item{x}{A vector of data}

\item{probs}{Probabilities to compute highest density regions for.}

\item{n}{Resolution of grid defined by range}

\item{range}{Range to compute regions.}

\item{parameters}{An optional list of parameters governing the density estimation procedure.}

\item{HDR_membership}{Should HDR membership of data points be calculated?
Defaults to \code{TRUE}, although it is computationally expensive.}

\item{fun}{Optional, a probability density function, must be vectorized in its first argument.}

\item{args}{Optional, a list of arguments to be provided to \code{fun}.}
}
\description{
TODO - fill in with how to to use \code{method} as a string, function factory, and how to use \code{fun}.
Likely will need several sections to describe various uses.
Also need to expand on \code{parameters} parameter.
}
