% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdensity-package.R
\docType{package}
\name{ggdensity}
\alias{ggdensity}
\alias{package-ggdensity}
\title{ggdensity: Stats and Geoms for Density Estimation with ggplot2}
\description{
A package that allows more flexible computations for visualization of density
estimates with ggplot2.
}
\seealso{
Useful links:
\itemize{
\item \url{https://jamesotto852.github.io/ggdensity}
\item \url{https://github.com/jamesotto852/ggdensity}
}
}
