% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_1d.R
\name{method_freqpoly_1d}
\alias{method_freqpoly_1d}
\title{Univariate frequency polygon HDR estimator}
\usage{
method_freqpoly_1d(bins = NULL)
}
\arguments{
\item{bins}{Number of bins. Defaults to normal reference rule (Scott, pg 59).}
}
\description{
temp
}
\details{
For more details on the use and implementation of the \verb{method_*_1d()} functions,
see \code{vignette("method", "ggdensity")}.
}
\references{
Scott, David W. Multivariate Density Estimation (2e), Wiley.
}
