% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_1d.R
\name{method_norm_1d}
\alias{method_norm_1d}
\title{Univariate parametric normal HDR estimator}
\usage{
method_norm_1d()
}
\description{
temp
}
\details{
For more details on the use and implementation of the \verb{method_*_1d()} functions,
see \code{vignette("method", "ggdensity")}.
}
