% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr_rug.R
\docType{data}
\name{geom_hdr_rug}
\alias{geom_hdr_rug}
\alias{stat_hdr_rug}
\alias{StatHdrRug}
\alias{GeomHdrRug}
\title{Rug plots of marginal highest density regions}
\format{
An object of class \code{StatHdrRug} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{GeomHdrRug} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
stat_hdr_rug(
  mapping = NULL,
  data = NULL,
  geom = "hdr_rug",
  position = "identity",
  ...,
  method = "kde",
  probs = c(0.99, 0.95, 0.8, 0.5),
  xlim = NULL,
  ylim = NULL,
  bw = "nrd0",
  adjust = 1,
  n = 512,
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)

StatHdrRug

geom_hdr_rug(
  mapping = NULL,
  data = NULL,
  stat = "hdr_rug",
  position = "identity",
  ...,
  outside = FALSE,
  sides = "bl",
  length = unit(0.03, "npc"),
  na.rm = FALSE,
  show.legend = TRUE,
  inherit.aes = TRUE
)

GeomHdrRug
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{method}{Density estimator to use, accepts character vector: \code{"kde"}}

\item{probs}{Probabilities to compute highest density regions for.}

\item{xlim, ylim}{Range to compute and draw regions. If \code{NULL}, defaults to
range of data.}

\item{adjust}{A multiplicative bandwidth adjustment to be used if \code{h} is
\code{NULL}.}

\item{n}{Resolution of grid used in discrete approximations for kernel
density and parametric estimators. Either a vector of length 2 or a scalar value
which is recycled for both dimensions.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{bins}{Number of bins along each axis for histogram and frequency polygon estimators.
Either a vector of length 2 or a scalar value which is recycled for both dimensions.
Defaults to normal reference rule (Scott, pg 87).}

\item{h}{Bandwidth for kernel density estimator. If \code{NULL}, estimated using
\code{\link[MASS:bandwidth.nrd]{MASS::bandwidth.nrd()}}}
}
\description{
Desc here
}
\section{Aesthetics}{
 geom_hdr_rug understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{x | y}
\item alpha
\item color
\item fill
\item group
\item linetype
\item size
\item subgroup
}
}

\section{Computed variables}{


\describe{ \item{level}{The level of the highest density region, specified
by \code{probs}, corresponding to each point.} }
}

\examples{

dfs <- tibble(
  x = rnorm(100),
  y = rnorm(100)
)

ggplot(dfs, aes(x)) +
  geom_density() +
  geom_hdr_rug(show.legend = FALSE) +
  coord_fixed()

ggplot(dfs, aes(y = y)) +
  geom_density() +
  geom_hdr_rug(show.legend = FALSE)

ggplot(dfs, aes(x, y)) +
  geom_point() +
  geom_hdr_rug(show.legend = FALSE) +
  coord_fixed()

ggplot(dfs, aes(x, y)) +
  geom_hdr() +
  geom_hdr_rug(show.legend = FALSE) +
  coord_fixed()



}
\references{
Scott, David W. Multivariate Density Estimation (2e), Wiley.
}
\keyword{datasets}
