% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hdr.R
\name{get_hdr}
\alias{get_hdr}
\title{Computing the highest density regions of a 2D density}
\usage{
get_hdr(
  data,
  method = "kde",
  probs = c(0.99, 0.95, 0.8, 0.5),
  n = 100,
  rangex = NULL,
  rangey = NULL,
  hdr_membership = TRUE,
  fun,
  args = list()
)
}
\arguments{
\item{data}{A data frame with columns \code{x} and \code{y}}

\item{method}{temp}

\item{probs}{Probabilities to compute highest density regions for.}

\item{n}{Resolution of grid defined by \code{xlim} and \code{ylim}.}

\item{rangex, rangey}{Range to compute and draw regions.}

\item{hdr_membership}{Should HDR membership of data points be calculated?
Defaults to \code{TRUE}, although it is computationally expensive.}

\item{fun}{Optional, a joint probability density function, must be vectorized in its first two arguments.}

\item{args}{Optional, a list of arguments to be provided to \code{fun}.}
}
\description{
TODO - fill in with how to to use \code{method} as a string, function factory, and how to use \code{fun}.
Likely will need several sections to describe various uses.
Also need to expand on \code{parameters} parameter.
}
